doing java on python
