+++ scripts/docs-collator/GitHubActionFetcher.py
@@ -1,22 +1,16 @@
 import os
 
+from AbstractFetcher import AbstractFetcher, MissingReadmeYamlException
+
 DOCS_DIR = 'docs'
 ACTIONS_DIR = 'actions'
-TARGETS_MD = 'targets.md'
 README_YAML = 'README.yaml'
 README_MD = 'README.md'
 
 
-class MissingReadmeYamlException(Exception):
-    def __init__(self):
-        self.message = "README.yaml is missing"
-        super().__init__(self.message)
-
-
-class GitHubActionFetcher:
+class GitHubActionFetcher(AbstractFetcher):
     def __init__(self, github_provider, download_dir):
-        self.download_dir = download_dir
-        self.github_provider = github_provider
+        super().__init__(github_provider, download_dir)
 
     def fetch(self, repo):
         repo_download_dir = os.path.join(self.download_dir, repo.name)
@@ -27,25 +21,13 @@ class GitHubActionFetcher:
         if README_YAML not in remote_files:
             raise MissingReadmeYamlException()
 
-        self.__fetch_readme_yaml(repo, repo_download_dir)
+        self._fetch_readme_yaml(repo, repo_download_dir)
 
         if DOCS_DIR in remote_files:
-            self.__fetch_docs(repo, repo_download_dir)
+            self._fetch_docs(repo, repo_download_dir)
 
         if ACTIONS_DIR in remote_files:
             self.__fetch_sub_actions(repo, repo_download_dir)
-
-    def __fetch_readme_yaml(self, repo, module_download_dir):
-        self.github_provider.fetch_file(repo, README_YAML, module_download_dir)
-
-    def __fetch_docs(self, repo, module_download_dir):
-        remote_files = self.github_provider.list_repo_dir(repo, DOCS_DIR)
-
-        for remote_file in remote_files:
-            if os.path.basename(remote_file) == TARGETS_MD:  # skip targets.md
-                continue
-
-            self.github_provider.fetch_file(repo, remote_file, module_download_dir)
 
     def __fetch_sub_actions(self, repo, module_download_dir):
         remote_files = self.github_provider.list_repo_dir(repo, ACTIONS_DIR)
