+++ scripts/docs-collator/render_docs_for_modules.py
@@ -3,10 +3,10 @@ import sys
 
 import click
 
+from AbstractRenderer import TerraformDocsRenderingError
 from GitHubProvider import GitHubProvider
 from ModuleFetcher import ModuleFetcher, MissingReadmeYamlException
 from ModuleRenderer import ModuleRenderer
-from Renderer import TerraformDocsRenderingError
 
 DOWNLOAD_TMP_DIR = 'tmp/modules'
 OUTPUT_DOC_DIR = 'content/modules/catalog'
